remove @inbounds for better performance?

+++ src/Hamiltonians/BoseHubbardMom1D2C.jl
@@ -138,7 +138,7 @@ function diagME(ham::BoseHubbardMom1D2C{T,HA,HB,V}, add::BoseFS2C{NA,NB,M,AA,AB}
     # ham_b = BoseHubbardMom1D(ham.nb, ham.m, ham.ub, ham.tb, add.bsb)
     onrep_a = BitStringAddresses.onr(add.bsa)
     onrep_b = BitStringAddresses.onr(add.bsb)
-    interaction2c = 0
+    interaction2c = Int32(0)
     for p in 1:M
         for k in 1:M
           interaction2c += onrep_a[k]*onrep_b[p] # b†_p b_p a†_k a_k
